Remove warnings

+++ intern/src/editor_gui/edit_inspector_light_probe.cpp
@@ -7,7 +7,7 @@ namespace Edit
 {
     CInspectorLightProbe::CInspectorLightProbe(QWidget* _pParent) 
         : QWidget          (_pParent)
-        , m_CurrentEntityID(-1)
+        , m_CurrentEntityID(static_cast<unsigned int>(-1))
     {
         // -----------------------------------------------------------------------------
         // Setup UI
